requirements.txt
folium
pandas
streamlit_folium
plotly.express
geopandas
arabic_reshaper
bidi.algorithm